+++ packages/allure-js-commons/src/framework/AllureConfig.ts
@@ -1,0 +1,7 @@
+import { TestResult } from "../model";
+import { AllureWriter } from "./AllureWriter";
+
+export interface AllureConfig {
+  readonly writer: AllureWriter;
+  readonly testMapper?: (test: TestResult) => TestResult | null;
+}
