+++ packages/allure-js-commons/src/framework/AllureGroup.ts
@@ -1,18 +1,22 @@
+import { TestResultContainer } from "../model";
+import { AllureExecutable } from "./AllureExecutable";
+import { fixtureResult, testResultContainer } from "./AllureResults";
 import { AllureRuntime } from "./AllureRuntime";
 import { AllureTest } from "./AllureTest";
-import { ExecutableItemWrapper } from "./ExecutableItemWrapper";
-import { fixtureResult, testResultContainer } from "./constructors";
-import { TestResultContainer } from "./model";
 
 export class AllureGroup {
-  private testResultContainer: TestResultContainer = testResultContainer();
+  private testResultContainer: TestResultContainer;
 
-  constructor(private readonly runtime: AllureRuntime) {}
+  constructor(private readonly runtime: AllureRuntime) {
+    this.testResultContainer = testResultContainer(runtime.crypto.uuid());
+  }
 
   startGroup(name?: string): AllureGroup {
     const group = new AllureGroup(this.runtime);
+
     this.testResultContainer.children.push(group.uuid);
     group.name = name || "Unnamed";
+
     return group;
   }
 
@@ -42,15 +46,17 @@ export class AllureGroup {
     this.testResultContainer.name = name;
   }
 
-  addBefore(): ExecutableItemWrapper {
+  addBefore(): AllureExecutable {
     const result = fixtureResult();
+
     this.testResultContainer.befores.push(result);
-    return new ExecutableItemWrapper(result);
+    return new AllureExecutable(result);
   }
 
-  addAfter(): ExecutableItemWrapper {
+  addAfter(): AllureExecutable {
     const result = fixtureResult();
+
     this.testResultContainer.afters.push(result);
-    return new ExecutableItemWrapper(result);
+    return new AllureExecutable(result);
   }
 }
