--- packages/allure-js-commons/src/isPromise.ts
@@ -1,2 +1,0 @@
-export const isPromise = (obj: any): boolean =>
-  !!obj && (typeof obj === "object" || typeof obj === "function") && typeof obj.then === "function";
