+++ packages/allure-js-commons/src/node/writers/InMemoryAllureWriter.ts
@@ -1,0 +1,8 @@
+import { PathLike, readFileSync } from "fs";
+import { AllureInMemoryAllureWriter as CommonInMemoryWriter } from "../../framework/writers/AllureInMemoryWriter";
+
+export class AllureInMemoryWriter extends CommonInMemoryWriter {
+  public writeAttachmentFromPath(from: PathLike, toFileName: string): void {
+    this.attachments[toFileName] = readFileSync(from);
+  }
+}
